Refactor connection types into separate files
