+++ src/bin/mavlink-dump.rs
@@ -15,7 +15,7 @@ fn main() {
     let args: Vec<_> = env::args().collect();
 
     if args.len() < 2 {
-        println!("Usage: mavlink-dump (tcpout|udpin|udpout|serial):(ip|dev):(port|baud)");
+        println!("Usage: mavlink-dump (tcpout|tcpin|udpout|udpin|serial):(ip|dev):(port|baud)");
         return;
     }
